fix: most react-kit circular dependencies
