+++ packages/react-kit/src/components/datepicker/types.ts
@@ -1,0 +1,5 @@
+export interface ChoosenTime {
+  hour: string | Array<string>;
+  minute: string | Array<string>;
+  timezone: string;
+}
