+++ packages/react-kit/src/components/exchangeCard/types.ts
@@ -1,0 +1,1 @@
+export type ExchangeCardStatus = "REDEEMED" | "CANCELLED" | "COMMITTED";
