+++ packages/react-kit/src/components/modal/ModalTypes.ts
@@ -1,0 +1,7 @@
+export const MODAL_TYPES = {
+  FINANCE_DEPOSIT_MODAL: "FINANCE_DEPOSIT_MODAL",
+  FINANCE_WITHDRAW_MODAL: "FINANCE_WITHDRAW_MODAL",
+  TRANSACTION_FAILED: "TRANSACTION_FAILED",
+  TRANSACTION_SUBMITTED: "TRANSACTION_SUBMITTED",
+  WAITING_FOR_CONFIRMATION: "WAITING_FOR_CONFIRMATION"
+} as const;
