+++ packages/react-kit/src/components/widgets/commit/CommitWidgetProviders.tsx
@@ -7,7 +7,7 @@ import {
   ConfigProviderProps
 } from "../../config/ConfigProvider";
 import { IpfsProvider, IpfsProviderProps } from "../../ipfs/IpfsProvider";
-import { MagicProvider } from "../../magicLink/MagicContext";
+import { MagicProvider } from "../../magicLink/MagicProvider";
 import ModalProvider from "../../modal/ModalProvider";
 import WalletConnectionProvider, {
   WalletConnectionProviderProps
