+++ packages/react-kit/src/hooks/callbacks/useRedemptionCallbacks.ts
@@ -1,57 +1,12 @@
-import { providers } from "ethers";
-import { FormType as RedeemFormType } from "../../components/modal/components/Redeem/RedeemFormModel";
 import { useRedemptionContext } from "../../components/widgets/redemption/provider/RedemptionContext";
-
-export type RedemptionInfo = {
-  exchangeId: string;
-  offerId: string;
-  buyerId: string;
-  sellerId: string;
-  sellerAddress: string;
-  buyerAddress: string;
-};
-
-export type DeliveryInfoMessage = RedemptionInfo & {
-  deliveryDetails: RedeemFormType;
-};
-
-export type RedeemTransactionMessage = {
-  isError: boolean;
-  error?: {
-    name: string;
-    message: string;
-  };
-  redemptionInfo: RedemptionInfo;
-};
-
-export type RedeemTransactionSubmittedMessage = RedeemTransactionMessage & {
-  redeemTx?: {
-    hash: string;
-  };
-};
-
-export type RedeemTransactionConfirmedMessage = RedeemTransactionMessage & {
-  redeemTx?: {
-    hash: string;
-    blockNumber: number;
-  };
-};
-
-export type DeliveryInfoCallbackResponse = {
-  accepted: boolean;
-  reason: string;
-  resume: boolean;
-};
-
-export type RedeemTransactionSubmittedCallbackResponse = {
-  accepted: boolean;
-  reason: string;
-};
-
-export type RedeemTransactionConfirmedCallbackResponse = {
-  accepted: boolean;
-  reason: string;
-};
+import {
+  DeliveryInfoCallbackResponse,
+  DeliveryInfoMessage,
+  RedeemTransactionConfirmedCallbackResponse,
+  RedeemTransactionConfirmedMessage,
+  RedeemTransactionSubmittedCallbackResponse,
+  RedeemTransactionSubmittedMessage
+} from "./types";
 
 async function fetchAndReadResponse(
   step: "deliveryInfo" | "redemptionSubmitted" | "redemptionConfirmed",
@@ -69,7 +24,9 @@ async function fetchAndReadResponse(
     let responseBody;
     try {
       responseBody = await response.json();
-    } catch {}
+    } catch {
+      //
+    }
     const accepted =
       response.ok && isTrueOrOkOrYes(responseBody?.accepted?.toString());
     if (!accepted) {
